Move methods into helperMethods.js, refactor Form
- Move the form for Edits into EditDataForm
