+++ src/utils/helperMethods.js
@@ -126,3 +126,31 @@ export function colourNode(node, id, index, searchTerm, colour, entryNode) {
     // e
   }
 }
+
+export function editFormat(edits) {
+  const editObj = {};
+  edits.map((edit) => {
+    editObj[edit.accessor] = edit.updated;
+  });
+  return editObj;
+}
+
+export function formEdits(original, updated) {
+  return Object.keys(updated).map((key) => {
+    if (original.vars[key] !== updated[key].data) {
+      return { accessor: updated[key].accessor, original: original.vars[key], updated: updated[key].data };
+    }
+    return null;
+  }).filter(a => a !== null);
+}
+
+export function hasFormChanged(original, updated) {
+  // We don't want to check all the data, just the data items that are editable
+  // Map over the data, return true/false based on whether the data has changed,
+  // then reduce using OR, so any true will return true.
+  return Object.keys(updated).map((key) => {
+    return original.vars[key] !== updated[key].data;
+  }).reduce((a, b) => {
+    return a || b;
+  }, false);
+}
